requirements.txt
torch
diffusers
transformers
accelerate
safetensors
pillow
gradio
